feat(test): add [HighMemory] attribute for 8GB+ RAM tests
- Added HighMemory category and attribute for tests requiring large allocations
- Updated LongIndexingMasterTest to use [HighMemory] (allocates 2.4GB arrays)
- Updated Broadcast_Copy_MaterializesFullArray to use [HighMemory]
- Updated CI workflow to exclude HighMemory tests along with OpenBugs

The [LongIndexing] category now runs in CI (smoke tests use 1MB, broadcast
tests use ~8 bytes). Only [HighMemory] tests are excluded from CI.

Filter syntax: '/*/*/*/*[Category!=OpenBugs]&/*/*/*/*[Category!=HighMemory]'

+++ test/NumSharp.UnitTest/OpenBugs.ApiAudit.cs
@@ -65,9 +65,10 @@ namespace NumSharp.UnitTest
             var result = a[mask];
 
             result.size.Should().Be(3, "Boolean mask selects 3 elements where mask is True");
-            result.GetInt32(0).Should().Be(0, "a[0] where mask[0]=True");
-            result.GetInt32(1).Should().Be(2, "a[2] where mask[2]=True");
-            result.GetInt32(2).Should().Be(4, "a[4] where mask[4]=True");
+            // arange returns Int64
+            result.GetInt64(0).Should().Be(0, "a[0] where mask[0]=True");
+            result.GetInt64(1).Should().Be(2, "a[2] where mask[2]=True");
+            result.GetInt64(2).Should().Be(4, "a[4] where mask[4]=True");
         }
 
         /// <summary>
